refactor(all): rafactor a lot

+++ src/five_in_a_row_game.cc
@@ -2,14 +2,14 @@
 // Created by small-sheep on 11/10/22.
 //
 
-#include "five_in_a_row_game/five_in_a_row_game.h"
+#include <five_in_a_row_game/five_in_a_row_game.h>
+#include <five_in_a_row_game/player.h>
+
 #include <iostream>
 #include <sstream>
 #include <string>
-#include "five_in_a_row_game/player.h"
 
-FiveInARowGame::FiveInARowGame() : board_pointer_(std::make_unique<Board>(5)) {
-}
+FiveInARowGame::FiveInARowGame() : board_pointer_(std::make_unique<Board>(5)) {}
 
 void FiveInARowGame::Start(Player *first_hand_player_ptr,
                            Player *second_hand_player_ptr) {
@@ -31,15 +31,15 @@ void FiveInARowGame::Tick() {
 }
 
 void FiveInARowGame::ProcessInput() {
-//  static std::string command;
-//  std::cout << "Please input your command:\n";
-//  std::cin >> command;
-//  ParseCommand(command);
+  //  static std::string command;
+  //  std::cout << "Please input your command:\n";
+  //  std::cin >> command;
+  //  ParseCommand(command);
 }
 
 void FiveInARowGame::Update() {
   last_frame_time_ = current_frame_time_;
-  current_frame_time_ = clock();
+  current_frame_time_ = static_cast<int>(clock());
   SetFramePerSecond(current_frame_time_ - last_frame_time_);
 
   first_hand_player_ptr_->ThinkAndPlaceAStone(board_pointer_.get());
@@ -48,12 +48,12 @@ void FiveInARowGame::Update() {
 
 void FiveInARowGame::Render() const {
   // Do some cleaning.
-//  std::cout << std::string(50, '\n');
+  //  std::cout << std::string(50, '\n');
   std::ostringstream ss;
   ss << FramePerSecond() << "\n";
   ss << "---- Map ----\n";
-  for (const auto &i: board_pointer_->BoardMap()) {
-    for (const auto &j: i) {
+  for (const auto &i : board_pointer_->BoardMap()) {
+    for (const auto &j : i) {
       ss << static_cast<int>(j) << " ";
     }
     ss << "\n";
@@ -67,17 +67,13 @@ void FiveInARowGame::ParseCommand(const std::string &command) {
   }
 }
 
-bool FiveInARowGame::GameIsOn() const {
-  return game_is_on_;
-}
+bool FiveInARowGame::GameIsOn() const { return game_is_on_; }
 
 void FiveInARowGame::SetGameIsOn(const bool game_is_on) {
   game_is_on_ = game_is_on;
 }
 
-int FiveInARowGame::FramePerSecond() const {
-  return frame_per_second_;
-}
+int FiveInARowGame::FramePerSecond() const { return frame_per_second_; }
 
 void FiveInARowGame::SetFramePerSecond(const int frame_per_second) {
   frame_per_second_ = frame_per_second;
